MAJOR UPDATE [READ DESCRIPTION]
Chat logging, uinfo, and so much more.. read under this as well, please.

The bot will start becoming more modular as it has went past v0.3. Through v0.3, we will be rewriting Precipitation to become more modular. This makes Precipitation easier to program, and will save a lot of trouble in the future. While normal Precipitation will be running, the new Precipitation will be running along side it, with the prefix pr;, so that it is still similar to the original. After the process is complete, it will go back to pr: but it will be modular instead!

Thank you, for taking the time to read this.